+++ CMakeLists.txt
@@ -20,45 +20,42 @@ project(libhal LANGUAGES CXX)
 
 libhal_unit_test(SOURCES
   tests/helpers.cpp
-  # tests/can.test.cpp
-  # tests/pwm.test.cpp
-  # tests/timer.test.cpp
-  # tests/i2c.test.cpp
-  # tests/spi.test.cpp
-  # tests/adc.test.cpp
-  # tests/dac.test.cpp
-  # tests/initializers.test.cpp
-  # tests/input_pin.test.cpp
-  # tests/interrupt_pin.test.cpp
-  # tests/output_pin.test.cpp
-  # tests/serial.test.cpp
-  # tests/steady_clock.test.cpp
-  # tests/motor.test.cpp
-  # tests/timeout.test.cpp
-  # tests/error.test.cpp
-  # tests/accelerometer.test.cpp
-  # tests/distance_sensor.test.cpp
-  # tests/gyroscope.test.cpp
-  # tests/magnetometer.test.cpp
-  # tests/rotation_sensor.test.cpp
-  # tests/temperature_sensor.test.cpp
-  # tests/servo.test.cpp
-  # tests/g_force.test.cpp
-  # tests/io_waiter.test.cpp
-  # tests/lengths.test.cpp
-  # tests/angular_velocity_sensor.test.cpp
-  # tests/current_sensor.test.cpp
-  # tests/stream_dac.test.cpp
-  # tests/lock.test.cpp
-  # tests/experimental/usb.test.cpp
-  # tests/zero_copy_serial.test.cpp
-  # tests/buffered_can.test.cpp
-  # tests/pointers.test.cpp
-  # tests/circular_buffer.test.cpp
-  # tests/coro_triple.test.cpp
-  # tests/coro_context.test.cpp
-  tests/coroutine.test.cpp
-  tests/coroutine.claude.test.cpp
+  tests/can.test.cpp
+  tests/pwm.test.cpp
+  tests/timer.test.cpp
+  tests/i2c.test.cpp
+  tests/spi.test.cpp
+  tests/adc.test.cpp
+  tests/dac.test.cpp
+  tests/initializers.test.cpp
+  tests/input_pin.test.cpp
+  tests/interrupt_pin.test.cpp
+  tests/output_pin.test.cpp
+  tests/serial.test.cpp
+  tests/steady_clock.test.cpp
+  tests/motor.test.cpp
+  tests/timeout.test.cpp
+  tests/error.test.cpp
+  tests/accelerometer.test.cpp
+  tests/distance_sensor.test.cpp
+  tests/gyroscope.test.cpp
+  tests/magnetometer.test.cpp
+  tests/rotation_sensor.test.cpp
+  tests/temperature_sensor.test.cpp
+  tests/servo.test.cpp
+  tests/g_force.test.cpp
+  tests/io_waiter.test.cpp
+  tests/lengths.test.cpp
+  tests/angular_velocity_sensor.test.cpp
+  tests/current_sensor.test.cpp
+  tests/stream_dac.test.cpp
+  tests/lock.test.cpp
+  tests/experimental/usb.test.cpp
+  tests/zero_copy_serial.test.cpp
+  tests/buffered_can.test.cpp
+  tests/pointers.test.cpp
+  tests/circular_buffer.test.cpp
+  tests/experimental/coroutine.test.cpp
   tests/main.test.cpp
 
   PACKAGES
