+++ tests/experimental/coroutine.test.cpp
@@ -1,16 +1,13 @@
-#include <array>
 #include <chrono>
-#include <coroutine>
 #include <cstddef>
 #include <memory_resource>
 #include <print>
-#include <span>
 #include <stdexcept>
 #include <string>
 #include <thread>
 #include <variant>
 
-#include <libhal/coroutine.hpp>
+#include <libhal/experimental/coroutine.hpp>
 
 #include <boost/ut.hpp>
